Update decision function and stock graph
I updated these files - when we switched to reading data directly from a URL some changes were needed due to the fact that the data is presented from newest to oldest previously and now from oldest to newest. Also updated the dataframe to store the date as a datetime object. Updated the graphing function to use the date as the X axis. Corrected the decision function (required due to change in CSV format).